refac+comments on remeshing
still needs polishing + question to cignoni about the valence refac & if
better to use setCrease, manually computing the crease
